pyproject.toml
[tool.poetry]
name = "mptt_assess"
version = "0.1.0"
description = ""
authors = ["Areiahna <areiahnacooks@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
